js/rm_load.js: add additional data search path
RPG Maker MV stores 'data' under 'www/data' not as a peer to the 'saves'
directory, so look in both locations.  This also adds a bit of additional
console output that was helpful in debugging and may be helpful if someone
is loading up their saves and some items seem to be missing.

Signed-off-by: Joe MacDonald <joe@deserted.net>

Co-authored-by: Nathan Baker <7409217+nathan-b@users.noreply.github.com>